% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{say_hello}
\alias{say_hello}
\title{A function to say hello to the user}
\usage{
say_hello(myname)
}
\arguments{
\item{myname}{The name of the person to say hello to.}
}
\value{
Says hello to the user input name.
}
\description{
A function to say hello to the user
}
\examples{
say_hello("Annie")
}
