% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_message}
\alias{send_message}
\title{Send a message}
\usage{
send_message(myname, nice = TRUE)
}
\arguments{
\item{myname}{Name to send the message to}

\item{nice}{A Boolean variable to indicate whether the message should be nice or mean.}
}
\value{
Either a nice message or a mean one.
}
\description{
Send a message
}
\examples{
send_message("Annie")
send_message("Annie", nice=FALSE)
}
